requirements.txt
Flask==2.3.2
flask-socketio==5.3.3
socketio-client==0.7.2
requests==2.31.0
